+++ node/src/BaseClient.ts
@@ -382,17 +382,17 @@ export class BaseClient {
     ) {
         const message = Array.isArray(command)
             ? redis_request.RedisRequest.create({
-                callbackIdx,
-                transaction: redis_request.Transaction.create({
-                    commands: command,
-                }),
-            })
+                  callbackIdx,
+                  transaction: redis_request.Transaction.create({
+                      commands: command,
+                  }),
+              })
             : command instanceof redis_request.Command
-                ? redis_request.RedisRequest.create({
+              ? redis_request.RedisRequest.create({
                     callbackIdx,
                     singleCommand: command,
                 })
-                : redis_request.RedisRequest.create({
+              : redis_request.RedisRequest.create({
                     callbackIdx,
                     scriptInvocation: command,
                 });
@@ -2103,9 +2103,9 @@ export class BaseClient {
         ReadFrom,
         connection_request.ReadFrom
     > = {
-            primary: connection_request.ReadFrom.Primary,
-            preferReplica: connection_request.ReadFrom.PreferReplica,
-        };
+        primary: connection_request.ReadFrom.Primary,
+        preferReplica: connection_request.ReadFrom.PreferReplica,
+    };
 
     /** Returns the element at index `index` in the list stored at `key`.
      * The index is zero-based, so 0 means the first element, 1 the second element and so on.
@@ -2237,11 +2237,11 @@ export class BaseClient {
             : undefined;
         const authenticationInfo =
             options.credentials !== undefined &&
-                "password" in options.credentials
+            "password" in options.credentials
                 ? {
-                    password: options.credentials.password,
-                    username: options.credentials.username,
-                }
+                      password: options.credentials.password,
+                      username: options.credentials.username,
+                  }
                 : undefined;
         const protocol = options.protocol as
             | connection_request.ProtocolVersion